Move ImageMap into game_map module
